Extract interfaces and objects into separate files
